Split up source control manager file to resolve SwiftLint error
